refactor(cli): extract evidence command family into cli/commands/evidence.ts
