+++ apps/axctl/src/cli/index.ts
@@ -120,6 +120,7 @@ import { renderCompareTable, renderCompareJson } from "./session-compare-format.
 import { cmdDaemon, cmdDoctor, cmdInstall, cmdSetup, cmdUninstall } from "./install.ts";
 import { insightsCommand, reportCommand, timelineCommand, reportRuntime } from "./commands/report.ts";
 import { signalsCommand, signalsRuntime } from "./commands/signals.ts";
+import { evidenceCommand, evidenceRuntime } from "./commands/evidence.ts";
 import type { RuntimeManifest } from "./commands/manifest.ts";
 import { resolvePwdRepository } from "../pwd.ts";
 import { detectStaleness } from "@ax/lib/transcript-staleness";
@@ -159,7 +160,6 @@ import {
     backtestEnforceWorktreeCase,
     formatFeedbackBacktestSummary,
 } from "../queries/feedback-cases.ts";
-import { guidanceNext, parseSelfImproveArgs, selfImproveWeekly, sessionSummary } from "../self-improve/commands.ts";
 import {
     buildIngestEventStatement,
     buildIngestRunStartStatement,
@@ -5164,45 +5164,6 @@ const hooksCommand = Command.make("hooks").pipe(
     ]),
 );
 
-const jsonSelfImprove = (cmd: "guidance" | "session" | "self-improve", rest: string[]) => {
-    const parsed = parseSelfImproveArgs(cmd, rest);
-    const effect =
-        parsed.command === "guidance-next" ? guidanceNext() :
-        parsed.command === "session-summary" ? sessionSummary() :
-        selfImproveWeekly();
-    return Effect.gen(function* () {
-        const result = yield* effect;
-        console.log(prettyPrint(result));
-    });
-};
-
-const evidenceGuidanceNextCommand = Command.make(
-    "guidance-next",
-    { json: jsonFlag },
-    ({ json }) => jsonSelfImprove("guidance", ["next", ...boolArg("json", json)]),
-).pipe(Command.withDescription("Return the next self-improvement guidance"));
-
-const evidenceSessionSummaryCommand = Command.make(
-    "session-summary",
-    { json: jsonFlag },
-    ({ json }) => jsonSelfImprove("session", ["summary", ...boolArg("json", json)]),
-).pipe(Command.withDescription("Summarize recent session evidence"));
-
-const evidenceWeeklyCommand = Command.make(
-    "weekly",
-    { json: jsonFlag },
-    ({ json }) => jsonSelfImprove("self-improve", ["weekly", ...boolArg("json", json)]),
-).pipe(Command.withDescription("Run weekly self-improvement evidence query"));
-
-const evidenceCommand = Command.make("evidence").pipe(
-    Command.withDescription("Self-improvement evidence queries (guidance, session, weekly)"),
-    Command.withSubcommands([
-        evidenceGuidanceNextCommand,
-        evidenceSessionSummaryCommand,
-        evidenceWeeklyCommand,
-    ]),
-);
-
 const bannerFlag = Flag.boolean("banner").pipe(Flag.withDefault(false));
 
 const versionCommand = Command.make(
@@ -5505,7 +5466,6 @@ const LEGACY_RUNTIME: RuntimeManifest = {
     hook: "db",
     hooks: "db",
     agents: "db",
-    evidence: "db",
     tui: "db",
     dogfood: "db",
 };
@@ -5514,6 +5474,7 @@ export const RUNTIME_BY_COMMAND: RuntimeManifest = {
     ...LEGACY_RUNTIME,
     ...reportRuntime,
     ...signalsRuntime,
+    ...evidenceRuntime,
 };
 
 // Commands whose handlers reach into SurrealClient via AppLayer (or the
